+++ src/misc/ExposeToJson.h
@@ -6,6 +6,6 @@ namespace ALZ {
 	class ParticleSystem;
 }
 
-#define EXPOSE_CUSTOMIZER_TO_JSON friend void toJson(nlohmann::json& j, const ParticleSystem& ps, int objectOrder);\
+#define EXPOSE_CUSTOMIZER_TO_JSON friend void toJson(nlohmann::json& j, const ParticleSystem& ps, size_t objectOrder);\
 								  friend void ParticleSystemFromJson(Environment* env, nlohmann::json& data, std::string id);\
 								  friend class ParticleSystem;
